backend/requirements.txt
bcrypt==3.2.0
bson==0.5.10
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.65.2
h11==0.12.0
importlib-metadata==4.6.0
jose==1.0.0
motor==2.4.0
pydantic==1.8.2
python-dotenv==0.19.0
python-jose==3.3.0
python-multipart==0.0.5
starlette==0.14.0
typing-extensions==3.10.0.0
uvicorn==0.14.0
passlib==1.7.4

